feature: move snippets and not found to app folder
